Fix MustNotBeNullWhiteSpace to throw ArgumentMustNotBeNullOrWhiteSpace (#15)
* Renamed `ArgumentWhitespaceException` and updated references

Renamed the class `ArgumentWhitespaceException` to `ArgumentWhiteSpaceException` and updated all references to this class.

Updated comments/documentation for the renamed method and class. Test methods in `ArgumentTests.cs` and `CommonThenSteps.cs` have been updated to expect the new exception type.

* Refactor `MustNotBeNullOrWhiteSpace` and add new exception class

Refactored the `MustNotBeNullOrWhiteSpace` method in `Argument.cs` to throw a new `ArgumentNullOrWhiteSpaceException` if the parameter is null or whitespace, replacing separate checks for null, empty, and whitespace.

Updated test methods in `ArgumentTests.cs` and `StringExtensionsTests.cs` to expect the new exception.

 Added `ArgumentNullOrWhiteSpaceException` class to `OpenMagic.Exceptions` namespace.

Introduced a new feature file `MustNotBeNullOrWhiteSpace.feature` describing the method's behavior in BDD style, and a corresponding steps class `MustNotBeNullOrWhiteSpaceSteps.cs`.

* Update OpenMagic project to version 5.2.0

Updated the version of the OpenMagic project from 5.1.0 to 5.2.0. This change is reflected in the `build-test-and-publish.yml`, `OpenMagic.csproj`, and `CommonAssemblyInfo.cs` files. The `VERSION` environment variable, `AssemblyVersion`, `FileVersion`, and `PackageVersion` have all been updated to 5.2.0.